Move Spring specific classes to module service-base-spring
The service-base module currently contains classes used for implementing
services and protocol adapters using Spring Boot.

In a first step towards splitting up the module into more specific
modules holding classes for implementing Spring/Quarkus based
services/protocol adapters, the Spring Boot specific application base
classes are moved to a separate service-base-spring module.
