+++ bom/pom.xml
@@ -159,6 +159,11 @@
       </dependency>
       <dependency>
         <groupId>org.eclipse.hono</groupId>
+        <artifactId>hono-service-base-spring</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>org.eclipse.hono</groupId>
         <artifactId>hono-service-device-registry-base</artifactId>
         <version>${project.version}</version>
       </dependency>
