+++ pom.xml
@@ -228,10 +228,10 @@
     <module>demo-certs</module>
     <module>deploy</module>
     <module>examples</module>
-    
     <module>legal</module>
     <module>service-base</module>
     <module>service-base-quarkus</module>
+    <module>service-base-spring</module>
     <module>services</module>
     <module>site</module>
     <module>test-utils</module>
