+++ services/command-router/src/main/java/org/eclipse/hono/commandrouter/Application.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 Contributors to the Eclipse Foundation
+ * Copyright (c) 2020, 2021 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -17,10 +17,10 @@ import java.util.Objects;
 
 import org.eclipse.hono.commandrouter.infinispan.EmbeddedCacheConfig;
 import org.eclipse.hono.commandrouter.infinispan.RemoteCacheConfig;
-import org.eclipse.hono.service.AbstractApplication;
 import org.eclipse.hono.service.AbstractServiceBase;
 import org.eclipse.hono.service.HealthCheckProvider;
 import org.eclipse.hono.service.auth.AuthenticationService;
+import org.eclipse.hono.service.spring.AbstractApplication;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
